Replaced standard ability link "flyouts" with custom flyouts for skills and spells.
Added a way to adjust points and levels on the skill/spell flyout windows. Allows for players to adjust the automatic calculation.
Reverted straight DB node copy for abilities. Now simply adds the properties needed to the character record.
Some fields on the skill/spell summary are readonly, including the autocalculated level.